ugly potential fix or break

+++ pages/teams/[teamId].js
@@ -47,7 +47,8 @@ import InfoTooltip from '../../components/InfoTooltip';
 //this function gathers the data from the player given in the path
 export async function getServerSideProps(context) {
     const { params } = context;
-    const teamId = parseInt(params.teamid);
+    const teamId = parseFloat(params.teamid);
+    //const teamId = parseInt(params.teamid);
 
     const prisma = new PrismaClient()
